+++ resources-ktx/CHANGELOG.md
@@ -3,3 +3,4 @@
 - **View**, **Fragment**: new extensions `getDimensionPixelSize` and `getDimensionPixelOffset`
 - **Context**: new extensions `resolveAttributeOrThrow` and `resolveAttribute`
 - **Context**: new extensions `resolveColor`
+- **Context**: new extension `resolveResourceId`